Added PyProbabilities
Let's say you are playing a game and there is an 8% chance of getting a weapon you want from killing a boss. This simple GUI written with tkinter allows you to enter .08 , and it tells you how many times you would need to kill the boss to have a 25%, 50%, 75%, and 99% chance of your loot dropping. 
ToDo: support more complex scenarios with multiple factors that effect the outcome